1. Added StatementDate on right Panel on Bank statement form.
2. Drag the Transaction on to the Line then the Transaction Amount will be
converted According to the Statement Line Date on Statement Line tab
in Statement window. And that converted amount
should be the Transaction amount on new form.

3. When drag the Trasaction on to the new form then the Transaction Amount
should convert according to the statementdate on new form and that converted
Amount should be as both txtTrxAmt and Amounts.

4. Drag Transaction on to the StatementLine, the StatementDate field should be
readonly on new form.

5. When change the Statement Date on new form the Amount should convert
according to ConversionRate of the that Statement Date.

6. TaxAmount for Tax Rate should calculate for Difference Amount
when the DifferenceType is Charge or if it is a Voucher type then TaxRate should
Calculate TaxAmount for Amount field.

7. ChargeAmount, Charge, TaxRate and Tax Amount fields on
StatementLine Tab in Statement window
should update when DifferenceType is Charge on new form.

8. Drag Prepay Order on to new form then the Amount field should readonly
it will create Payment with full Amount.

9. Drag Contra on to the new form then the Converted(if currency is not same) amount should update
on TxtAmt and Amount fields.
--> if Conversion Rate not found it will show popUp message.